[NFC] Refactoring the methods for removing dead allocs and reshape folding into interface patterns.
This adds `populate*` methods to get patterns to remove dead
allocations and to fold reshapes into `hal.interface.binding.subspan`
so that they could be used in different passes.
